Use awaitSyncAndIBD, mv awaitSyncAndIBD to NodeTestUtil
